Add task 2.4
