+++ app/src/main/java/otus/gpb/homework/activities/ActivityD.kt
@@ -2,11 +2,10 @@ package otus.gpb.homework.activities
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 
-class ActivityD : AppCompatActivity() {
+class ActivityD : AbstractAppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
