+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,7 @@
 <resources>
     <string name="app_name">Activities</string>
     <string name="open_activity_b">Open ActivityB</string>
+    <string name="open_activity_d">Open ActivityD</string>
     <string name="open_activity_c">Open ActivityC</string>
     <string name="open_activity_a">Open ActivityA</string>
     <string name="close_activity_c">Close ActivityC</string>
